Do NOT use Text instead of String
I used String because it is the type used by the optparse-applicative
and typed-process libraries, suspecting that using Text would not
improve performance.  I did a test with my real configuration file,
setting the command to echo and using the last entry, and measuring the
real runtime using the time command.  Both the version using String and
the version using Text ran in 0.013 seconds.  Since there is no
measurable difference, I am keeping the String version because it at
least avoids lots of pack/unpack calls.